[Workplace Search] Factor out Document-level Permissions components for re-use (#120131)
* Factor out DocumentPermissionsField component

* Factor out DocumentPermissionsCallout component
