+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/content_sources/components/add_source/connect_instance.test.tsx
@@ -12,11 +12,10 @@ import React from 'react';
 
 import { shallow } from 'enzyme';
 
-import { EuiSwitch } from '@elastic/eui';
-
 import { staticSourceData } from '../../source_data';
 
 import { ConnectInstance } from './connect_instance';
+import { DocumentPermissionsCallout } from './document_permissions_callout';
 
 describe('ConnectInstance', () => {
   // Needed to mock redirect window.location.replace(oauthUrl)
@@ -126,13 +125,6 @@ describe('ConnectInstance', () => {
     expect(setSourceSubdomainValue).toHaveBeenCalledWith(TEXT);
   });
 
-  it('calls handler on click', () => {
-    const wrapper = shallow(<ConnectInstance {...props} />);
-    wrapper.find(EuiSwitch).simulate('change', { target: { checked: true } });
-
-    expect(setSourceIndexPermissionsValue).toHaveBeenCalledWith(true);
-  });
-
   it('handles form submission with oauth source', () => {
     jest.spyOn(window.location, 'replace').mockImplementationOnce(mockReplace);
     const wrapper = shallow(<ConnectInstance {...oauthProps} />);
@@ -145,23 +137,10 @@ describe('ConnectInstance', () => {
     expect(mockReplace).toHaveBeenCalled();
   });
 
-  it('renders doc-level permissions message when not available', () => {
-    const wrapper = shallow(<ConnectInstance {...props} needsPermissions={false} />);
-
-    expect(wrapper.find('FormattedMessage')).toHaveLength(1);
-  });
-
-  it('renders callout when not synced', () => {
-    setMockValues({ ...values, indexPermissionsValue: false });
-    const wrapper = shallow(<ConnectInstance {...props} />);
-
-    expect(wrapper.find('EuiCallOut')).toHaveLength(1);
-  });
-
   it('renders documentLevelPermissionsCallout', () => {
     setMockValues({ ...values, hasPlatinumLicense: false });
     const wrapper = shallow(<ConnectInstance {...oauthProps} />);
 
-    expect(wrapper.find('[data-test-subj="DocumentLevelPermissionsCallout"]')).toHaveLength(1);
+    expect(wrapper.find(DocumentPermissionsCallout)).toHaveLength(1);
   });
 });
